edit traits
