+++ gears/src/x/keepers/bank.rs
@@ -5,11 +5,8 @@ use crate::{
     context::{QueryableContext, TransactionalContext},
     error::AppError,
     types::{
-        address::AccAddress,
-        base::{coin::Coin, send::SendCoins},
-        denom::Denom,
-        store::gas::errors::GasStoreErrors,
-        tx::metadata::Metadata,
+        address::AccAddress, base::send::SendCoins, denom::Denom,
+        store::gas::errors::GasStoreErrors, tx::metadata::Metadata,
     },
     x::module::Module,
 };
@@ -37,23 +34,10 @@ pub trait BankKeeper<SK: StoreKey, M: Module>: Clone + Send + Sync + 'static {
         base: &Denom,
     ) -> Result<Option<Metadata>, GasStoreErrors>;
 
-    fn balance_all<DB: Database, CTX: QueryableContext<DB, SK>>(
-        &self,
-        ctx: &CTX,
-        address: &AccAddress,
-    ) -> Result<Vec<Coin>, GasStoreErrors>;
-
     fn coins_burn<DB: Database, CTX: TransactionalContext<DB, SK>>(
         &self,
         ctx: &mut CTX,
         module: &M,
         deposit: &SendCoins,
     ) -> Result<(), AppError>;
-
-    fn balance<DB: Database, CTX: QueryableContext<DB, SK>>(
-        &self,
-        ctx: &CTX,
-        address: &AccAddress,
-        denom: &Denom,
-    ) -> Result<Coin, GasStoreErrors>;
 }
