+++ gears/src/x/keepers/mod.rs
@@ -1,3 +1,4 @@
+pub mod gov;
 pub mod staking;
 pub mod auth;
 pub mod bank;
